autoencoder-dimensionality-reduction/requirements.txt
numpy==1.24.3  # Numerical computing for autoencoder training
pyyaml==6.0.1  # YAML configuration file parsing
python-dotenv==1.0.0  # Environment variable management
pytest==7.4.0  # Testing framework
pytest-cov==4.1.0  # Test coverage reporting

